rename command to dbctl
